==> 4_create_dataset.ipynb <==
# Create Dataset

Here we have to decide how to split the dataset to create training/val/test sets.  
We decided to:
1. Create the graph with all the reviews before 2021 using reviews as nodes.
2. Then the dropped reviews are divided into positive (1) if they have 4 or 5 stars and negative (0) if they have 1 or 2 stars:
- reviews in 2021 form training and validation sets (80/20)
- reviews in 2022 form test set


import pandas as pd
import torch
import dgl

review_df = pd.read_csv('neo4j_csvs/review.csv')

review_ids = pd.read_csv('preprocessed/review_ids.csv')
business_ids = pd.read_csv('preprocessed/business_ids.csv')
user_ids = pd.read_csv('preprocessed/user_ids.csv')

review_df['date'] = pd.to_datetime(review_df['date'], infer_datetime_format=True)  

review_df

train_pos = review_df[(review_df['date'].dt.year == 2021) & (review_df['stars:float'] > 3)]
train_neg = review_df[(review_df['date'].dt.year == 2021) & (review_df['stars:float'] < 3)]
test_pos = review_df[(review_df['date'].dt.year == 2022) & (review_df['stars:float'] > 3)]
test_neg = review_df[(review_df['date'].dt.year == 2022) & (review_df['stars:float'] < 3)]

train_pos = torch.tensor(review_ids[review_ids['review_id:ID'].isin(train_pos['review_id:ID'])].index.values)
train_neg = torch.tensor(review_ids[review_ids['review_id:ID'].isin(train_neg['review_id:ID'])].index.values)

test_pos = torch.tensor(review_ids[review_ids['review_id:ID'].isin(test_pos['review_id:ID'])].index.values)
test_neg = torch.tensor(review_ids[review_ids['review_id:ID'].isin(test_neg['review_id:ID'])].index.values)

graph = dgl.load_graphs('preprocessed/graph.dgl')
graph = graph[0][0]

train_business_pos = graph.out_edges(train_pos, etype='review_to_business')[1]
train_business_neg = graph.out_edges(train_neg, etype='review_to_business')[1]
test_business_pos = graph.out_edges(test_pos, etype='review_to_business')[1]
test_business_neg = graph.out_edges(test_neg, etype='review_to_business')[1]

train_user_pos = graph.in_edges(train_pos, etype='user_to_review')[0]
train_user_neg = graph.in_edges(train_neg, etype='user_to_review')[0]
test_user_pos = graph.in_edges(test_pos, etype='user_to_review')[0]
test_user_neg = graph.in_edges(test_neg, etype='user_to_review')[0]

gen1 = torch.Generator().manual_seed(42)
gen2 = torch.Generator().manual_seed(42)
split_pos = torch.utils.data.random_split(range(train_user_pos.shape[0]), [0.8, 0.2], generator=gen1)
split_neg = torch.utils.data.random_split(range(train_user_neg.shape[0]), [0.8, 0.2], generator=gen2)

train_user_pos_final = torch.index_select(train_user_pos, 0, torch.tensor(split_pos[0]))
val_user_pos_final = torch.index_select(train_user_pos, 0, torch.tensor(split_pos[1]))

train_business_pos_final = torch.index_select(train_business_pos, 0, torch.tensor(split_pos[0]))
val_business_pos_final = torch.index_select(train_business_pos, 0, torch.tensor(split_pos[1]))

train_user_neg_final = torch.index_select(train_user_neg, 0, torch.tensor(split_neg[0]))
val_user_neg_final = torch.index_select(train_user_neg, 0, torch.tensor(split_neg[1]))

train_business_neg_final = torch.index_select(train_business_neg, 0, torch.tensor(split_neg[0]))
val_business_neg_final = torch.index_select(train_business_neg, 0, torch.tensor(split_neg[1]))

train_pos = (train_user_pos_final, train_business_pos_final)
train_neg = (train_user_neg_final, train_business_neg_final)

val_pos = (val_user_pos_final, val_business_pos_final)
val_neg = (val_user_neg_final, val_business_neg_final)

test_pos = (test_user_pos, test_business_pos)
test_neg = (test_user_neg, test_business_neg)

graph = dgl.remove_nodes(graph, torch.tensor(review_ids[review_ids['review_id:ID'].isin(review_df[review_df['date'].dt.year >= 2021]['review_id:ID'])].index.values), ntype='review')

graph.num_nodes

import pickle
dgl.save_graphs('training_data/graph.dgl', [graph])

with open('training_data/train.obj', 'wb') as fp:
	pickle.dump({'pos': train_pos, 'neg': train_neg}, fp)

with open('training_data/val.obj', 'wb') as fp:
	pickle.dump({'pos': val_pos, 'neg': val_neg}, fp)

with open('training_data/test.obj', 'wb') as fp:
	pickle.dump({'pos': test_pos, 'neg': test_neg}, fp)
