==> 6_inference.ipynb <==
# Inference

```
docker-compose up 
```

## connection

from qdrant_client import QdrantClient
from qdrant_client.http.models import Distance, VectorParams

client = QdrantClient("localhost", port=6333)
client.recreate_collection(
    collection_name="business",
    vectors_config=VectorParams(size=128, distance=Distance.DOT),
)

client.get_collection(collection_name="business")

## Populate collection

import pandas as pd
import dgl
import torch

# we can do it also as call on Neo4j
business_ids = pd.read_csv('preprocessed/business_ids.csv')
category_ids = pd.read_csv('preprocessed/category_ids.csv')
category_df = pd.read_csv('neo4j_csvs/categories.csv')
category_business_rels = pd.read_csv('neo4j_csvs/category_business_rels.csv')
review_business_rels = pd.read_csv('neo4j_csvs/review_business_rels.csv')
user_review_rels = pd.read_csv('neo4j_csvs/user_review_rels.csv')

graph = dgl.load_graphs('training_data/graph.dgl')
graph = graph[0][0]
edges = {}
for canonical_etype in graph.canonical_etypes:
    edges[canonical_etype] = graph.edges(etype=canonical_etype)

# we add the reversed relations
edges[('category', 'category_to_business', 'business')] = (graph.edges(etype='business_has_category')[1], graph.edges(etype='business_has_category')[0])
edges[('business', 'business_to_review', 'review')] = (graph.edges(etype='review_to_business')[1], graph.edges(etype='review_to_business')[0])
edges[('business', 'business_to_tip', 'tip')] = (graph.edges(etype='tip_to_business')[1], graph.edges(etype='tip_to_business')[0])
edges[('review', 'review_to_user', 'user')] = (graph.edges(etype='user_to_review')[1], graph.edges(etype='user_to_review')[0])
edges[('tip', 'tip_to_user', 'user')] = (graph.edges(etype='user_to_tip')[1], graph.edges(etype='user_to_tip')[0])

num_nodes_dict = {} 
for ntype in graph.ntypes:
    num_nodes_dict[ntype] = graph.nodes(ntype).shape[0]

g = dgl.heterograph(edges, num_nodes_dict = num_nodes_dict)
g.ndata['feat'] = {k: torch.tensor(v, dtype=torch.float32) for k, v in graph.ndata['feat'].items() }
del graph

business_ids # ordered as business in graph dgl

category_business_rels # business -> category

from utils import HGT
node_dict = { ntype: g.ntypes.index(ntype) for ntype in g.ntypes }
edge_dict = { canonical_etype: g.canonical_etypes.index(canonical_etype) for canonical_etype in g.canonical_etypes }
feature_dim_dict = { ntype: g.ndata['feat'][ntype].shape[1] for ntype in g.ntypes }
model = HGT(node_dict, edge_dict, feature_dim_dict, n_hid=256, n_out=128, n_layers=4, n_heads=8, use_norm=False)
sampler = dgl.dataloading.NeighborSampler([24, 24, 24, 24])

model.load_state_dict(torch.load('models/best_model.pt')['model_state_dict'])

from qdrant_client.http.models import PointStruct
from qdrant_client.http.models import UpdateStatus

batch_size = 64
model.eval()
def split(df, chunk_size):
    for i in range(df.index.start, df.index.stop, chunk_size):
        yield list(range(i, i + chunk_size if df.index.stop > i + chunk_size else df.index.stop)), df.loc[i:i + chunk_size - 1]

for batch in split(business_ids, batch_size):
    with torch.no_grad():
        pos_batch_business_ids, batch_business_ids = batch

        pos_block_business = [blocks for _, _, blocks in dgl.dataloading.DataLoader(
            g, { 'business': pos_batch_business_ids }, sampler,
            batch_size=len(pos_batch_business_ids), shuffle=False, drop_last=False, num_workers=1)][0]
        pos_business_logits = model(pos_block_business, 'business')
        batch_categories = batch_business_ids.merge(category_business_rels, left_on='business_id:ID', right_on=':START_ID')
        businesses_its_categories = batch_categories.groupby(by=["business_id:ID"], sort=False)[':END_ID'].apply(list).reset_index()
        
        batch_business_ids = batch_business_ids['business_id:ID'].to_list()
        businesses_its_categories = businesses_its_categories[':END_ID'].to_list()
        pos_business_logits = pos_business_logits.tolist()
        points = []
        for i in range(len(pos_batch_business_ids)):
            points.append(
                PointStruct(id=pos_batch_business_ids[i], vector=pos_business_logits[i], payload={'id': batch_business_ids[i], 'categories': businesses_its_categories[i]})
            )
        operation_info = client.upsert(
            collection_name="business",
            wait=True,
            points=points
        )
        assert operation_info.status == UpdateStatus.COMPLETED


from qdrant_client import QdrantClient

client = QdrantClient("localhost", port=6333, timeout=100)

client.create_snapshot(collection_name='business')

client.get_collection(collection_name="business")

## Recommendations

user_ids = pd.read_csv('preprocessed/user_ids.csv')

input_user_id = 'MGPQVLsODMm9ZtYQW-g_OA'
input_category_id = '386c1f850fbd5f478fb4ef8a134c1740'

pos_user_id = user_ids[user_ids['user_id:ID'] == input_user_id].index[0] # getting DGL id to create user embedding

user_ids

already_reviewed_businesses = user_ids[user_ids['user_id:ID'] == input_user_id] \
    .merge(user_review_rels, left_on='user_id:ID', right_on=':START_ID') \
    .merge(review_business_rels, left_on=':END_ID', right_on=':START_ID', suffixes=('rev', 'business'))[':END_IDbusiness'].to_list()

len(already_reviewed_businesses)

pos_block_user = [blocks for _, _, blocks in dgl.dataloading.DataLoader(
    g, {'user': [pos_user_id] }, sampler,
    batch_size=batch_size, shuffle=False, drop_last=False, num_workers=1)][0]
with torch.no_grad():
    pos_user_logits = model(pos_block_user, 'user')

from qdrant_client.http.models import Filter, FieldCondition, MatchValue, MatchAny


search_result = client.search(
    collection_name="business",
    query_vector=pos_user_logits.tolist()[0], 
    query_filter=Filter(
        must=[
            FieldCondition(
                key="categories",
                match=MatchValue(value=input_category_id)
            )
        ],
        must_not=[
            FieldCondition(
                key="id", 
                match=MatchAny(any=already_reviewed_businesses)
            ),
        ],
    ),
    limit=100
)

search_result
